Add MessageService
Move some function in IssueService to MessageService
